dev FrontEndFunctionSearch
