+++ FrontEnd/App.tsx
@@ -22,13 +22,11 @@ import RatingSucessfullyScreen from './android/app/src/Screens/Raiting/RaitingSu
 import EditRatingScreen from './android/app/src/Screens/Raiting/EditRating';
 import LoginScreen from './android/app/src/Screens/Login/Login';
 import BookScreen from './android/app/src/Screens/Books/OrderIsOnGoing';
-
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-
-
 import { OrderTrackingScreen } from './android/app/src/Screens/Books/OrderTracking';
 import { OrderDetail } from './android/app/src/Screens/Books/OrderDetail';
 import { ServicePage } from './android/app/src/Screens/AddOrder/Service';
+import SearchScreen from './android/app/src/Screens/Home/Search';
 
 const Stack = createNativeStackNavigator();
 
@@ -158,6 +156,11 @@ function App(): React.JSX.Element {
           <Stack.Screen
             name="OrderDetail"
             component={OrderDetail}
+            options={{headerShown: false}}
+          />
+           <Stack.Screen
+            name="SearchScreen"
+            component={SearchScreen}
             options={{headerShown: false}}
           />
         <Stack.Screen name="HomeScreen" component={Homestack} />
