+++ FrontEnd/android/app/src/Hook/useRegisterUser.ts
@@ -72,7 +72,7 @@ const useRegisterUser = (): RegisterHook => {
       password,
     };
     axios
-      .post('https://b936-2402-9d80-458-dfa5-1cd6-169-fc7b-e541.ngrok-free.app/api/Register', user)
+      .post('https://b9b7-14-176-231-248.ngrok-free.app/api/Register', user)
       .then(response => {
         console.log('User registered:', response.data);
         Alert.alert('Đăng ký thành công');
@@ -84,7 +84,7 @@ const useRegisterUser = (): RegisterHook => {
         console.error('Đăng ký thất bại:', error);
         Alert.alert('Đăng ký không thành công');
       });
-  }
+  };
   return {
     username,
     phone,
