+++ FrontEnd/android/app/src/Screens/Home/components/Order.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { Image,StyleSheet, Text, TouchableOpacity, View } from 'react-native';
-
+import { Image, Text, TouchableOpacity, View } from 'react-native';
+import styles from '../../../Styles/Order';
 export const renderOrder = ({ item }: any) => {
   return (
     <View style={styles.itemContainer}>
@@ -28,43 +28,3 @@ export const renderOrder = ({ item }: any) => {
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  Iconwashing: {
-    width: 20,
-    height: 20,
-  },
-  iconContainer: {
-    borderColor: '#7EB9C1',
-    borderWidth: 2.5,
-    justifyContent: 'center',
-    padding: 10,
-    margin: 5,
-    borderRadius: 50,
-  },
-  itemContainer: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    backgroundColor: '#5A617B',
-    borderRadius: 10,
-    alignItems: 'center',
-   padding:2,
-   marginRight:8
-  },
-  IdOrder: {
-    color: 'white',
-  },
-  reBuyButton:{
-    backgroundColor:'#7EB9C1',
-    marginTop:20,
-    borderRadius:10
-  },
-  title:{
-    fontWeight:'bold',
-    padding:10,
-    color:'white'
-  },
-  containerButton:{
-   
-  }
-});
