+++ FrontEnd/android/app/src/Styles/HomePage.ts
@@ -1,0 +1,9 @@
+import { StyleSheet } from "react-native";
+const styles = StyleSheet.create({
+    container: {
+      backgroundColor: '#353B51',
+      height: '100%',
+    },
+    header: {},
+  });
+export  default styles;
