+++ FrontEnd/android/app/src/Styles/NearLocation.ts
@@ -1,0 +1,40 @@
+import { StyleSheet } from "react-native";
+const styles = StyleSheet.create({
+    viewitemNotification: {
+      backgroundColor: '#5A617B',
+      padding: 15,
+      borderRadius: 15,
+      flexDirection: 'row',
+      marginStart: 20,
+      marginRight: 10,
+    },
+    viewImgae: {
+      width: 70,
+      height: 50,
+      objectFit:'contain'
+    },
+    textdescription: {
+      color: '#BBF4FB',
+      fontWeight: '600',
+    },
+    textseedetail: {
+      color: '#FFF',
+      fontWeight: '600',
+      fontSize: 11,
+      marginTop: 4,
+    },
+    textTime: {
+      color: '#BBF4FB',
+      fontWeight: '600',
+      fontSize: 11,
+      marginTop: 1,
+    },
+    viewStatus: {
+      backgroundColor: '#BBF4FB',
+      borderRadius: 50,
+      width: 10,
+      height: 10,
+      marginTop: 4,
+    },
+  });
+  export  default styles;
