+++ FrontEnd/android/app/src/Styles/RenderMiniGame.ts
@@ -1,0 +1,19 @@
+import { StyleSheet } from "react-native";
+const styles = StyleSheet.create({
+    container:{
+        flexDirection:'row',
+        marginLeft:20,
+        marginTop:10,
+    },
+    image:{
+        width:50,
+        height:50,
+        borderRadius:10
+    },
+    name:{
+        color:'white',
+        fontSize:10,
+        width:'60%'
+    }
+});
+export  default styles;
