+++ FrontEnd/android/app/src/Styles/Search.ts
@@ -1,0 +1,58 @@
+import { StyleSheet } from "react-native";
+const styles = StyleSheet.create({
+    viewsearch:{
+        flex:1,
+        backgroundColor:'#353B51',
+    },
+  view:{
+      padding:20, 
+      flexDirection:'row',
+      marginRight:25,
+    },
+    viewSearch:{
+      borderRadius:15,
+      borderWidth:1,
+      backgroundColor:'#fff',
+      borderColor:'#FFF',
+      paddingLeft:40,
+      flexDirection:'row',
+      width:'100%',
+    },
+    textInputSearch:{
+      paddingRight:50,
+    },
+    viewSearches:{
+      paddingLeft:20,
+      paddingRight:20,
+    },
+    viewbranch:{
+      flexDirection:'column',
+      gap:20,
+    },
+    textall:{
+       color:'#fff',
+       fontSize:18,
+       fontWeight:'600',
+    },
+    viewlocation:{
+      backgroundColor:'#5A617B',
+      borderRadius:50,
+     padding:10,
+    },
+    showItem:{
+      flexDirection:"row",
+      gap:20,
+    },
+    viewborder:{
+      borderColor:'#fff',
+      borderWidth:0.5,
+    }
+    ,viewAddress:{
+      marginTop:5,
+    },
+    textname:{
+      color:'#fff',
+      fontSize:14,
+    }
+  });
+export default styles;
