+++ FrontEnd/android/app/src/Styles/SlideShowImage.ts
@@ -1,0 +1,19 @@
+import { StyleSheet } from "react-native";
+const styles = StyleSheet.create({
+    imageSlide: {
+      width: 350,
+      height: 200,
+      borderRadius:20,
+      objectFit:'fill'
+    },
+    SlideShow: {
+      marginLeft:30,
+      marginRight:30,
+      marginBottom:30,
+      alignItems:'center',
+    },
+    FlastlistSlide:{
+      width:350,
+    }
+  });
+ export default styles;
